+++ src/utils/calculations.ts
@@ -187,3 +187,4 @@ export function validateSkillValue(skillValue: number, masteryRank: number): num
 
 
 
+